# docs/openapi/openapi.yaml
openapi: 3.0.4
info:
  title: OrgFlow API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /requests:
    post:
      operationId: createRequest
      tags:
        - Requests
      summary: request作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestInput'
      responses:
        '201':
          description: request投稿成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
    get:
      operationId: listRequests
      tags:
        - Requests
      summary: request を一覧表示する
      parameters:
        - name: scope
          description: request 一覧の絞り込み条件。省略時は全件、pending-approval 指定時は承認対象のみ返す
          in: query
          required: false
          schema:
            type: string
            enum:
              - pending-approval
      responses:
        '200':
          description: request 一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRequestsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateRequestInput:
      type: object
      required:
        - title
        - internalOrganizationId
        - requestType
        - amount
      properties:
        title:
          type: string
          maxLength: 100
        internalOrganizationId:
          type: string
          format: uuid
        requestType:
          type: string
          enum: [transportation_expenses, travel_expenses, equipment_purchase_expenses]
        amount:
          type: integer
          minimum: 0
...
    Request:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: リクエストID
        tenantId:
          type: string
          format: uuid
          description: テナントID
        requestSeriesId:
          type: string
          format: uuid
          description: リクエストシリーズID（同一申請の改版をまとめる単位）
        versionNo:
          type: integer
          description: バージョン番号
        status:
          type: string
          enum: [draft, in_progress, approved, returned, rejected, cancelled]
          description: リクエストの状態
        currentAppliedApprovalRouteStepId:
          type: string
          format: uuid
          nullable: true
          description: 現在の承認ルートステップID
        applicantUserId:
          type: string
          format: uuid
          description: 申請者ユーザーID
        internalOrganizationId:
          type: string
          format: uuid
          description: 申請元の内部組織ID
        requestType:
          type: string
          enum: [transportation_expenses, travel_expenses, equipment_purchase_expenses]
          description: 申請種別
        title:
          type: string
          maxLength: 100
          description: 申請の件名
        description:
          type: string
          nullable: true
          description: 申請の詳細
        amount:
          type: integer
          minimum: 0
          description: 申請金額（円）
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
...
    ListRequestsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Request'
...
  responses:
    UnauthorizedError:
      description: 認証に失敗しました
    ForbiddenError:
      description: 権限が不足しています
...
    UnprocessableEntityError:
      description: 入力値が不正です